+++ packages/uniforms-bootstrap3/src/NumField.tsx
@@ -8,7 +8,7 @@ const noneIfNaN = (x: any) => (isNaN(x) ? undefined : x);
 const parse = (decimal: any, x: any) =>
   noneIfNaN((decimal ? parseFloat : parseInt)(x));
 
-const Num_ = (props: any) =>
+const Num = (props: any) =>
   wrapField(
     props,
     <input
@@ -20,7 +20,11 @@ const Num_ = (props: any) =>
       max={props.max}
       min={props.min}
       name={props.name}
-      onChange={props.onChange}
+      onChange={event => {
+        const parse = props.decimal ? parseFloat : parseInt;
+        const value = parse(event.target.value);
+        props.onChange(isNaN(value) ? undefined : value);
+      }}
       placeholder={props.placeholder}
       ref={props.inputRef}
       step={props.step || (props.decimal ? 0.01 : 1)}
@@ -28,47 +32,5 @@ const Num_ = (props: any) =>
       value={props.value}
     />,
   );
-
-let Num;
-// istanbul ignore next
-if (parseInt(React.version, 10) < 16) {
-  Num = class Num extends Component<any, any> {
-    state = { value: '' + this.props.value };
-
-    componentWillReceiveProps({ decimal, value }: any) {
-      if (
-        parse(decimal, value) !==
-        parse(decimal, this.state.value.replace(/[.,]+$/, ''))
-      ) {
-        this.setState({
-          value: value === undefined || value === '' ? '' : '' + value,
-        });
-      }
-    }
-
-    onChange = (event: any) => {
-      const value = event.target.value.replace(/[^\d.,-]/g, '');
-
-      this.setState({ value });
-      this.props.onChange(parse(this.props.decimal, value));
-    };
-
-    render() {
-      return Num_({
-        ...this.props,
-        onChange: this.onChange,
-        value: this.state.value,
-      });
-    }
-  };
-} else {
-  Num = (props: any) =>
-    Num_({
-      ...props,
-      onChange(event: any) {
-        props.onChange(parse(props.decimal, event.target.value));
-      },
-    });
-}
 
 export default connectField(Num);