+++ packages/uniforms-material/src/NumField.tsx
@@ -2,10 +2,7 @@ import React, { Component } from 'react';
 import TextField from '@material-ui/core/TextField';
 import { connectField, filterDOMProps } from 'uniforms';
 
-const noneIfNaN = x => (isNaN(x) ? undefined : x);
-const parse = (decimal, x) => noneIfNaN((decimal ? parseFloat : parseInt)(x));
-
-const Num_ = ({
+const Num = ({
   decimal,
   disabled,
   error,
@@ -30,7 +27,11 @@ const Num_ = ({
     inputProps={{ min, max, step: decimal ? 0.01 : 1, ...inputProps }}
     label={label}
     name={name}
-    onChange={onChange}
+    onChange={event => {
+      const parse = decimal ? parseFloat : parseInt;
+      const value = parse(event.target.value);
+      onChange(isNaN(value) ? undefined : value);
+    }}
     placeholder={placeholder}
     ref={inputRef}
     type="number"
@@ -38,48 +39,6 @@ const Num_ = ({
     {...filterDOMProps(props)}
   />
 );
-
-let Num;
-// istanbul ignore next
-if (parseInt(React.version, 10) < 16) {
-  Num = class Num extends Component<any, any> {
-    state = { value: '' + this.props.value };
-
-    componentWillReceiveProps({ decimal, value }) {
-      if (
-        parse(decimal, value) !==
-        parse(decimal, this.state.value.replace(/[.,]+$/, ''))
-      ) {
-        this.setState({
-          value: value === undefined || value === '' ? '' : '' + value,
-        });
-      }
-    }
-
-    onChange = event => {
-      const value = event.target.value.replace(/[^\d.,-]/g, '');
-
-      this.setState({ value });
-      this.props.onChange(parse(this.props.decimal, value));
-    };
-
-    render() {
-      return Num_({
-        ...this.props,
-        onChange: this.onChange,
-        value: this.state.value,
-      } as any);
-    }
-  };
-} else {
-  Num = props =>
-    Num_({
-      ...props,
-      onChange(event) {
-        props.onChange(parse(props.decimal, event.target.value));
-      },
-    });
-}
 
 Num.defaultProps = {
   fullWidth: true,