+++ packages/uniforms-semantic/src/NumField.tsx
@@ -2,10 +2,7 @@ import React, { Component } from 'react';
 import classnames from 'classnames';
 import { connectField, filterDOMProps } from 'uniforms';
 
-const noneIfNaN = x => (isNaN(x) ? undefined : x);
-const parse = (decimal, x) => noneIfNaN((decimal ? parseFloat : parseInt)(x));
-
-const Num_ = ({
+const Num = ({
   className,
   decimal,
   disabled,
@@ -49,7 +46,11 @@ const Num_ = ({
         max={max}
         min={min}
         name={name}
-        onChange={onChange}
+        onChange={event => {
+          const parse = decimal ? parseFloat : parseInt;
+          const value = parse(event.target.value);
+          onChange(isNaN(value) ? undefined : value);
+        }}
         placeholder={placeholder}
         ref={inputRef}
         step={step || (decimal ? 0.01 : 1)}
@@ -67,47 +68,5 @@ const Num_ = ({
     )}
   </div>
 );
-
-let Num;
-// istanbul ignore next
-if (parseInt(React.version, 10) < 16) {
-  Num = class Num extends Component<any, any> {
-    state = { value: '' + this.props.value };
-
-    componentWillReceiveProps({ decimal, value }) {
-      if (
-        parse(decimal, value) !==
-        parse(decimal, this.state.value.replace(/[.,]+$/, ''))
-      ) {
-        this.setState({
-          value: value === undefined || value === '' ? '' : '' + value,
-        });
-      }
-    }
-
-    onChange = event => {
-      const value = event.target.value.replace(/[^\d.,-]/g, '');
-
-      this.setState({ value });
-      this.props.onChange(parse(this.props.decimal, value));
-    };
-
-    render() {
-      return Num_({
-        ...this.props,
-        onChange: this.onChange,
-        value: this.state.value,
-      } as any);
-    }
-  };
-} else {
-  Num = props =>
-    Num_({
-      ...props,
-      onChange(event) {
-        props.onChange(parse(props.decimal, event.target.value));
-      },
-    });
-}
 
 export default connectField(Num);